separate socket and server
