+++ Runtime/Scripts/Constant.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Runtime.InteropServices;
 using System.Text;
 using UnityEditor;
 
@@ -8,7 +7,7 @@ namespace OscCore
 #if UNITY_EDITOR
     [InitializeOnLoad]
 #endif
-    public static class Constant
+    static class Constant
     {
         public const byte Comma = (byte) ',';
         public const byte ForwardSlash = (byte) '/';
@@ -22,7 +21,6 @@ namespace OscCore
             bundleBytes[7] = 0;
             BundlePrefixBytes = bundleBytes;
             BundlePrefixLong = BitConverter.ToInt64(bundleBytes, 0);
-            var handle = GCHandle.Alloc(BundlePrefixBytes, GCHandleType.Pinned);
         }
     }
 }
